requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
requests==2.32.3
sqlalchemy==2.0.36
alembic==1.14.0
pydantic==2.9.2
psycopg[binary]
python-dotenv
